services/core/requirements.txt
alembic==1.3.0
aniso8601==8.0.0
bcrypt==3.1.7
blinker==1.4
boto3==1.10.27
botocore==1.13.27
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
coverage==4.5.3
DateTime==4.3
docutils==0.15.2
entrypoints==0.3
flake8==3.7.7
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-DebugToolbar==0.10.1
Flask-Migrate==2.4.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.2.1
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
reportlab==3.5.32
requests==2.21.0
s3transfer==0.2.1
six==1.13.0
SQLAlchemy==1.3.10
urllib3==1.24.3
Werkzeug==0.16.0
zope.interface==4.6.0
